vals

+++ feb19/class-practice/index.js
@@ -44,12 +44,24 @@ function drawJokes(arr) {
 
     const favButton = card.querySelector(`#fav-${joke.id}`);
     favButton.addEventListener("click", function () {
-      favs.push(joke);
-      localStorage.setItem("favJokes", JSON.stringify(favs));
+      if (favs.some((item) => item.id === joke.id)) {
+        Swal.fire({
+          icon: "error",
+          title: "Oops...",
+          text: "This joke is already in the favorites!",
+        });
+      } else {
+        favs.push(joke);
+        localStorage.setItem("favJokes", JSON.stringify(favs));
 
-      const newArr = arr;
-      newArr.splice(i, 1);
-      drawJokes(newArr);
+        const newArr = arr;
+        newArr.splice(i, 1);
+        // drawJokes(newArr);
+        Swal.fire({
+          title: "Added to favorites!",
+          icon: "success",
+        });
+      }
     });
 
     const delButton = card.querySelector(`#del-${joke.id}`);
@@ -60,6 +72,10 @@ function drawJokes(arr) {
       if (strFavs.includes(`${joke.id}`)) {
         localStorage.setItem("favJokes", JSON.stringify(arr));
       }
+      Swal.fire({
+        title: "Joke removed successfully!",
+        icon: "success",
+      });
     });
 
     GLOBAL.masterContainer.appendChild(card);
